% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_email.r
\name{build_email}
\alias{build_email}
\title{Builds and sends the email}
\usage{
build_email(from_, to_, creds, body_, table_iss, table_pr, footer_)
}
\arguments{
\item{from_}{sender email address}

\item{to_}{recipient email address}

\item{creds}{A credential file of the type supported by code{blastula}}

\item{body_}{The text to precede the table}

\item{table_iss}{the issue table}

\item{table_pr}{The PR table}

\item{footer_}{the text to follow the table}
}
\value{
the email body
}
\description{
This function compiles the tables with some email text and sends it
using \code{blastula::smtp_send.}
}
\seealso{
\link[blastula:credential_helpers]{blastula::creds}
}
