% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_.R
\name{calculate_cumulative_stars}
\alias{calculate_cumulative_stars}
\title{Calculate the cumulative number of stars}
\usage{
calculate_cumulative_stars(data)
}
\arguments{
\item{data}{A data frame returned from \code{\link[=get_stargazers]{get_stargazers()}}.}
}
\value{
A tibble with the following six columns is returned:
\itemize{
\item org: A string, specifying the GitHub organization.
\item repo: A string, specifying the GitHub repository.
\item date: A date in the format of \code{yyyy-mm-dd}.
\item day: The number of days from the initial star as an integer.
\item stars: The number of stars that were added on that day.
\item cumulative_stars: The number of total stars since the first day the
repository was starred.
}
}
\description{
For each combination of organization and repository, calculate the
cumulative number of stars from the date of the first star. Where, stars
represent when a GitHub user "stars" a repository.
}
