% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_issue_table.R
\name{create_issue_table}
\alias{create_issue_table}
\title{Creates a table of issues}
\usage{
create_issue_table(json_input)
}
\arguments{
\item{json_input}{A \code{json} table generated from the Github API.}
}
\value{
A \code{html} table of issues.
}
\description{
This function takes a \code{json} input table generated from the Github
API, and creates an issue table with the fields number, title, labels,
milestone, pull_request, and assignee. The table is returned as an
\code{html} table grouped by milestone.
}
