% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pr_table.R
\name{create_pr_table}
\alias{create_pr_table}
\title{Creates a table of pull requests}
\usage{
create_pr_table(json_input)
}
\arguments{
\item{json_input}{A \code{json} table generated from the Github API.}
}
\value{
A \code{html} table of pull requests.
}
\description{
This function takes a \code{json} input table generated from the Github
API, and creates a pull request table with the fields number, title, labels,
milestone, pull_request, and assignee. The table is returned as an
\code{html} table grouped by milestone.
}
