% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_issues}
\alias{get_issues}
\title{Query the Github API for a list of issues associated with a repository}
\usage{
get_issues(org = "NOAA-FIMS", repo = "FIMS")
}
\arguments{
\item{org}{The name of the organization that contains the \code{repo}. This can
also be a user name rather than an organization. The default is
\code{"NOAA-FIMS"}.}

\item{repo}{The name of the GitHub repository. The default is \code{"FIMS"}.}
}
\value{
A 33-column tibble, that was created from the \code{json} list that is returned
from \code{\link[gh:gh]{gh::gh()}}, is returned. The returned tibble contains the following
columns:
\itemize{
\item url
\item repository_url
\item labels_url
\item comments_url
\item events_url
\item html_url
\item id
\item node_id
\item number
\item title
\item user
\item labels
\item state
\item locked
\item assignee
\item assignees
\item milestone
\item comments
\item created_at
\item updated_at
\item closed_at
\item author_association
\item sub_issues_summary
\item active_lock_reason
\item draft
\item pull_request: A five-column data frame with information about the Pull
Request if the issue was a actually a Pull Request and not an issue. All
entries are NULL if it was in fact an issue. Columns include \code{url},
\code{html_url}, \code{diff_url}, \code{patch_url}, and \code{merged_at}.
\item body
\item closed_by
\item reactions
\item timeline_url
\item performed_via_github_app
\item state_reason
}
}
\description{
The issues for a given repository and all of their metadata are returned as
a tibble. The table can then be used to summarize information about a given
repository such as how many open issues there are. Information about closed
issues is not returned.
}
\examples{
\dontrun{
get_issues(repo = "FIMS", org = "NOAA-FIMS")
}
}
