% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repositories_.r
\name{get_repositories_user}
\alias{get_repositories_user}
\title{Get a list of repositories for a given user}
\usage{
get_repositories_user(user)
}
\arguments{
\item{user}{A string providing the GitHub user name of interest.}
}
\value{
A tibble is returned with the following three columns:
\itemize{
\item user: A string with the provided GitHub user name.
\item name: A string giving the name of the repository.
\item fork: A logical, specifying if the repository is a fork.
}
}
\description{
Get a list of repositories for a given user
}
\examples{
get_repositories_user("kellijohnson-NOAA")
data <- purrr::map_df(
  c("kellijohnson-NOAA", "msupernaw"),
  get_repositories_user
)
}
