% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_star_history_.R
\name{get_star_history_repo}
\alias{get_star_history_repo}
\title{Get the history of stars for a repository}
\usage{
get_star_history_repo(org, repo)
}
\arguments{
\item{org}{A string providing the name of the GitHub organization that the
\code{repo} is located in.}

\item{repo}{A string providing the name of the GitHub repository that you
are interested in getting information for.}
}
\value{
A tibble with the following four columns is returned:
\itemize{
\item date: A date in the format of \code{yyyy-mm-dd}.
\item day: The number of days from the initial star as an integer.
\item stars: The number of stars that were added on that day.
\item cumulative_stars: The number of total stars since the first day the
repository was starred.
}
}
\description{
Get the history of when a repository was starred and return it as a tibble.
This information is useful for plotting.
}
\examples{
stars <- get_star_history_repo("NOAA-FIMS", "FIMS")
plot(
  x = stars[["day"]],
  y = stars[["cumulative_stars"]],
  type = "l",
  xlab = "Day since first star",
  ylab = "Cumulative stars"
)
}
