% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issues.R
\name{get_issues}
\alias{get_issues}
\title{Query the Github API}
\usage{
get_issues(appname_, key_, secret_, org_name, repo_name)
}
\arguments{
\item{appname_}{The name of the application you have authorized on Github.}

\item{key_}{The key you have authorized on Github.}

\item{secret_}{The secret corresponding to the key.}

\item{org_name}{The organization or username the repo is stored under.}

\item{repo_name}{The name of the repository}
}
\value{
A \code{json} list of returned issues from Github.
}
\description{
This function takes as input the Github API app name, the Github API
key, the corresponding secret, the organization and repo name, and uses
the Github api and \code{httr} package to query the API for the issues
in that repository.
}
