% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issues.R
\name{get_issues}
\alias{get_issues}
\title{Query the Github API}
\usage{
get_issues(appname_, key_, secret_, repo_name, type = "issues")
}
\arguments{
\item{appname_}{The name of the application you have authorized on Github.}

\item{key_}{The key you have authorized on Github.}

\item{secret_}{The secret corresponding to the key.}

\item{repo_name}{The name of the repository}

\item{type}{The API endpoint you are querying}
}
\value{
A \code{json} list of returned issues from Github.
}
\description{
This function takes as input the Github API app name, the Github API
key, the corresponding secret, the organization and repo name, and uses
the Github api and \code{httr} package to query the API for the issues
in that repository.
}
\examples{
\dontrun{
get_issues(appname_ = "github",
key_ = "56b637a5baffac62cad9",
secret_ = "8e107541ae1791259e9987d544ca568633da2ebf",
repo_name = "r-lib/httr")
}
}
