% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repositories_.r
\name{get_repositories}
\alias{get_repositories}
\title{Get a list of repositories for a given user or organization}
\usage{
get_repositories(x, type = c("users", "orgs"))
}
\arguments{
\item{user}{A string providing the GitHub user name or organization name of
interest.}
}
\value{
A tibble is returned with the following three columns:
\itemize{
\item user: A string with the provided GitHub user name or organization.
\item name: A string giving the name of the repository.
\item fork: A logical, specifying if the repository is a fork.
}
}
\description{
Get a list of repositories for a given user or organization
}
\examples{
get_repositories("kellijohnson-NOAA")
data <- purrr::map_df(
  c("kellijohnson-NOAA", "msupernaw"),
  get_repositories
)
FIMS_repos <- get_repositories("NOAA-FIMS", "orgs")
}
