% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_.R
\name{get_stargazers}
\alias{get_stargazers}
\title{Get the history of stars for a repository}
\usage{
get_stargazers(org = "NOAA-FIMS", repo = "FIMS", from = -Inf)
}
\arguments{
\item{org}{The name of the organization that contains the \code{repo}. This can
also be a user name rather than an organization. The default is
\code{"NOAA-FIMS"}.}

\item{repo}{The name of the GitHub repository. The default is \code{"FIMS"}.}

\item{from}{The date that you would like to start the activity from. The
default is negative infinity, which leads to all entries being returned.
But, another viable option is \code{Sys.Date() - 1}, which provides everything
since yesterday. Note that some entries prior to this entry will also be
returned if they are included in the same page, as the entire page is
returned.}
}
\value{
A tibble with 24 columns is returned. The first twenty columns are from the
JSON data that is returned from the GitHub API and the next four columns are
added internally, specifying the GitHub organization (org), the GitHub
repository (repo), the date the star was added, and the days since the first
star was added to the repository compared when this star was added (day).
}
\description{
Get the history of when a repository was starred and return it as a tibble.
}
\examples{
stars <- get_stargazers("NOAA-FIMS", "FIMS")
plot(
  x = stars[["day"]],
  y = stars[["cumulative_stars"]],
  type = "l",
  xlab = "Day since first star",
  ylab = "Cumulative stars"
)
}
